Create README.md
The project was "Visitor pass" in which when a visitor comes to visit the campus they will show their Aadhar card or pen card in the camera and security guard will capture it and after capturing, information from the card will be extracted and filled in the form after which security ask for approval from admin and admin got a notification whether to accept the visitor request or reject and all the information of a visitor will be saved in the database.

There are two types of users 1) Admin and 2) Security. Admin can add security and can accept or reject visitors and manage security and visitors. Security will capture the card photo and send the details to Admin and after getting approval from the admin, he will allow the visitor to visit the campus, and when a visitor leaves, he will checkout that visitor.